feat: categorize dropdown
