+++ src/components/BankTransactionsLoader/BankTransactionsLoader.tsx
@@ -1,9 +1,8 @@
-
 import ChevronDownFill from '../../icons/ChevronDownFill'
 import { BankTransaction } from '../../types'
 import { IconButton, SubmitButton } from '../Button'
 import { SubmitAction } from '../Button/SubmitButton'
-import { CategorySelect } from '../CategorySelect'
+import { CategorySelect } from '../CategorySelect/CategorySelect'
 import { SkeletonTableLoader } from '../SkeletonTableLoader'
 
 export const BankTransactionsLoader = ({
@@ -47,13 +46,13 @@ export const BankTransactionsLoader = ({
       cols={
         isLoading
           ? [
-              { colSpan: 5, trimLastXRows: 4 },
-              { colSpan: 1, parts: 2 },
-            ]
+            { colSpan: 5, trimLastXRows: 4 },
+            { colSpan: 1, parts: 2 },
+          ]
           : [
-              { colSpan: 4 },
-              { colSpan: 1, colComponent: inactiveBankTransactionsActions },
-            ]
+            { colSpan: 4 },
+            { colSpan: 1, colComponent: inactiveBankTransactionsActions },
+          ]
       }
       height={20}
     />
