+++ src/components/CategorySelect/CategorySelect.tsx
@@ -1,28 +1,18 @@
-import Select, {
-  DropdownIndicatorProps,
-  GroupHeadingProps,
-  GroupBase,
-  OptionProps,
-  components,
-} from 'react-select'
-import { DATE_FORMAT } from '../../config/general'
-import { useLayerContext } from '../../contexts/LayerContext'
-import Check from '../../icons/Check'
-import ChevronDown from '../../icons/ChevronDown'
-import InfoIcon from '../../icons/InfoIcon'
-import MinimizeTwo from '../../icons/MinimizeTwo'
-import { centsToDollars as formatMoney } from '../../models/Money'
-import { BankTransaction, CategorizationType, Category } from '../../types'
-import { SuggestedMatch, type CategoryWithEntries } from '../../types/bank_transactions'
-import { Badge } from '../Badge'
-import { BadgeSize } from '../Badge/Badge'
-import { Tooltip, TooltipContent, TooltipTrigger } from '../Tooltip'
-import { Text, TextSize } from '../Typography'
+import { Button, Popover, DialogTrigger, Dialog, ListBox, ListBoxSection, Header, ListBoxItem, Collection } from 'react-aria-components'
+import { Input } from '../Input/Input'
+import { ReactNode, useMemo, useState } from 'react'
+import { useLayerContext } from '../../contexts/LayerContext/LayerContext'
+import { BankTransaction, Category } from '../../types'
+import { CategoryOption, CategoryOptionPayload, CategoryWithHide } from './types'
+import { buildMatchOptions, buildSuggestedOptions, filterCategories } from './utils'
+import { Text, TextSize, TextWeight } from '../Typography'
+import Plus from '../../icons/Plus'
 import { CategorySelectDrawer } from './CategorySelectDrawer'
 import classNames from 'classnames'
-import { parseISO, format as formatTime } from 'date-fns'
+import pluralize from 'pluralize'
+import ChevronDown from '../../icons/ChevronDown'
 
-type Props = {
+type CategorySelectProps = {
   name?: string
   bankTransaction: BankTransaction
   value: CategoryOption | undefined
@@ -34,217 +24,80 @@ type Props = {
   asDrawer?: boolean
 }
 
-export enum OptionActionType {
-  CATEGORY = 'category',
-  MATCH = 'match',
-  HIDDEN = 'hidden',
+const MAX_INDENT_LEVEL = 3
+
+const MenuSection = ({ children }: { children: ReactNode }) => (
+  <div className='Layer__category-select__menu-section'>{children}</div>
+)
+
+const MenuItem = ({ text, icon }: { text: string, icon?: ReactNode }) => (
+  <div className='Layer__category-select__menu-item'>
+    <Text>{text}</Text>
+    {icon}
+  </div>
+)
+
+type CategoryListProps = {
+  option: Category & { hide?: boolean }
+  level: number
+  accountName: string
+  onSelect: (newValue: CategoryOption) => void
+  selected: CategoryOption | undefined
 }
 
-export interface CategoryOptionPayload {
-  id: string
-  option_type: OptionActionType
-  display_name: string
-  description?: string
-  date?: string
-  amount?: number
-  type?: string
-  stable_name?: string
-  entries?: CategoryWithEntries['entries']
-  subCategories: Category[] | null
-}
-
-export interface CategoryOption {
-  type: string
-  disabled?: boolean
-  payload: CategoryOptionPayload
-}
-
-export const mapCategoryToOption = (category: CategoryWithEntries): CategoryOption => {
-  return {
-    type: OptionActionType.CATEGORY,
-    payload: {
-      id: 'id' in category ? category.id : '',
-      option_type: OptionActionType.CATEGORY,
-      display_name: category.display_name,
-      type: category.type,
-      description: category.description ?? undefined,
-      stable_name: ('stable_name' in category) ? category.stable_name ?? '' : '',
-      entries: category.entries,
-      subCategories: category.subCategories,
-    },
-  }
-}
-
-export const mapCategoryToExclusionOption = (
-  category: CategoryWithEntries & { type: 'ExclusionNested' },
-): CategoryOption => {
-  return {
-    type: OptionActionType.CATEGORY,
-    payload: {
-      id: category.id,
-      option_type: OptionActionType.CATEGORY,
-      display_name: category.display_name,
-      type: 'ExclusionNested',
-      stable_name: '',
-      entries: category.entries,
-      subCategories: category.subCategories,
-    },
-  }
-}
-
-export const mapSuggestedMatchToOption = (
-  record: SuggestedMatch,
-): CategoryOption => {
-  return {
-    type: OptionActionType.MATCH,
-    payload: {
-      id: record.id,
-      option_type: OptionActionType.MATCH,
-      display_name: record.details.description,
-      amount: record.details.amount,
-      subCategories: null,
-    },
-  }
-}
-
-const DropdownIndicator:
-  | React.ComponentType<
-    DropdownIndicatorProps<CategoryOption, false, GroupBase<CategoryOption>>
-  >
-  | null
-  | undefined = (props) => {
+const CategoryList = ({ option, level = 0, accountName, onSelect, selected }: CategoryListProps) => {
+  if (option?.subCategories) {
     return (
-      <components.DropdownIndicator {...props}>
-        <ChevronDown />
-      </components.DropdownIndicator>
-    )
-  }
-
-const GroupHeading = (
-  props: GroupHeadingProps<CategoryOption, false, GroupBase<CategoryOption>>,
-) => {
-  return (
-    <components.GroupHeading
-      className={classNames(
-        props.className,
-        props.children === 'Match' || props.children === 'All categories'
-          ? 'Layer__select__group-heading--main'
-          : '',
-      )}
-      {...props}
-    />
-  )
-}
-
-const Option = (
-  props: OptionProps<CategoryOption, false, GroupBase<CategoryOption>> & {
-    showTooltips: boolean
-  },
-) => {
-  if (props.data.payload.option_type === 'hidden') {
-    return null
-  }
-
-  if (props.data.type === 'match') {
-    return (
-      <components.Option
-        {...props}
-        className={`${props.className} Layer__select__option-content__match`}
-      >
-        <div className='Layer__select__option-content__match__main-row'>
-          <span className='Layer__select__option-content__match__date'>
-            {props.data.payload.date
-            && formatTime(parseISO(props.data.payload.date), DATE_FORMAT)}
-          </span>
-          <span className='Layer__select__option-content__match__description'>
-            {props.data.payload.display_name}
-          </span>
-        </div>
-        <div className='Layer__select__option-content__match__amount-row'>
-          <span className='Layer__select__option-content__match__amount'>
-            $
-            {formatMoney(props.data.payload.amount)}
-          </span>
-        </div>
-      </components.Option>
+      <>
+        <Header
+          style={{
+            display: option.hide ? 'none' : 'flex',
+            paddingLeft: `${(Math.min(MAX_INDENT_LEVEL, Math.max(level, 1)) * 12) + 8}px`,
+          }}
+        >
+          {level === 0
+            ? (
+              <Text size={TextSize.xs} status='disabled'>{option.display_name}</Text>
+            )
+            : (
+              <Text weight={TextWeight.bold}>{option.display_name}</Text>
+            )}
+        </Header>
+        <ListBoxSection className='Layer__category-select__list-box-section' key={`${accountName}-${option.category}-section`} style={{ display: (option as CategoryWithHide).hide ? 'none' : 'flex' }}>
+          {option.subCategories.map((o, i) => (
+            <Collection key={`${accountName}-${option.category}-section-${i}`} items={o.subCategories ?? []}>
+              <CategoryList option={o} level={level + 1} accountName={option.display_name} onSelect={onSelect} selected={selected} />
+            </Collection>
+          ))}
+        </ListBoxSection>
+      </>
     )
   }
 
   return (
-    <components.Option
-      {...props}
-      className={`Layer__select__option-menu-content ${props.className}`}
+    <ListBoxItem
+      key={`${accountName}-${option.category}`}
+      className='Layer__category-select__list-item'
+      style={{
+        paddingLeft: `${8 + (Math.min(MAX_INDENT_LEVEL, level) * 12)}px`,
+        flexDirection: 'row',
+        justifyContent: 'space-between',
+        alignItems: 'center',
+        width: '100%',
+        display: option.hide ? 'none' : 'flex',
+      }}
+      textValue={option?.display_name}
+      onAction={() => {
+        onSelect({
+          type: 'category',
+          payload: { ...option, option_type: 'category' } as CategoryOptionPayload,
+        })
+      }}
     >
-      <div className='Layer__select__option-menu--name'>
-        {props.isSelected
-          ? (
-            <span className='Layer__select__option-menu-content-check'>
-              <Check size={16} />
-            </span>
-          )
-          : (
-            <span className='Layer__select__option-menu-content-check'>
-              <div style={{ width: 16, height: 16 }} />
-            </span>
-          )}
-        <div>{props.data.payload.display_name}</div>
-      </div>
-      {props.showTooltips && props.data.payload.description && (
-        <div className='Layer__select__option-menu--tooltip'>
-          <Tooltip>
-            <TooltipTrigger>
-              <InfoIcon />
-            </TooltipTrigger>
-            <TooltipContent className='Layer__actionable-list__tooltip-content'>
-              <Text
-                className='Layer__actionable-list__content-description'
-                size={TextSize.sm}
-              >
-                {props.data.payload.description}
-              </Text>
-            </TooltipContent>
-          </Tooltip>
-        </div>
-      )}
-    </components.Option>
+      <Text>{option?.display_name}</Text>
+      <Text size={TextSize.sm} status='disabled'>{accountName}</Text>
+    </ListBoxItem>
   )
-}
-
-const allCategoriesDivider: GroupBase<CategoryOption>[] = [
-  {
-    label: 'All categories',
-    options: [
-      {
-        type: 'All categories',
-        disabled: true,
-        payload: {
-          id: 'all_categories',
-          option_type: OptionActionType.HIDDEN,
-          display_name: 'ALL CATEGORIES',
-          subCategories: null,
-        },
-      } satisfies CategoryOption,
-    ],
-  },
-]
-
-function flattenCategories(
-  categories: Category[],
-): GroupBase<CategoryOption>[] {
-  function getLeafCategories(category: Category): Category[] {
-    if (!category.subCategories || category.subCategories.length === 0) {
-      return [category]
-    }
-    return category.subCategories.flatMap(subCategory =>
-      getLeafCategories(subCategory),
-    )
-  }
-  return categories.map((category) => {
-    return {
-      label: category.display_name,
-      options: getLeafCategories(category).map(x => mapCategoryToOption(x)),
-    } satisfies GroupBase<CategoryOption>
-  })
 }
 
 export const CategorySelect = ({
@@ -257,65 +110,28 @@ export const CategorySelect = ({
   showTooltips,
   excludeMatches = false,
   asDrawer = false,
-}: Props) => {
+}: CategorySelectProps) => {
+  const [isOpen, setIsOpen] = useState(false)
+  const [search, setSearch] = useState('')
+
   const { categories } = useLayerContext()
 
-  const matchOptions =
-    !excludeMatches && bankTransaction?.suggested_matches
-      ? [
-          {
-            label: 'Match',
-            options: bankTransaction.suggested_matches.map((x) => {
-              return {
-                type: OptionActionType.MATCH,
-                payload: {
-                  id: x.id,
-                  option_type: OptionActionType.MATCH,
-                  display_name: x.details.description,
-                  date: x.details.date,
-                  amount: x.details.amount,
-                  subCategories: null,
-                },
-              } satisfies CategoryOption
-            }),
-          } satisfies GroupBase<CategoryOption>,
-      ]
-      : []
+  const matches = buildMatchOptions(bankTransaction, excludeMatches, search)
+  const suggestions = buildSuggestedOptions(bankTransaction, search)
+  const allCategories = useMemo<CategoryWithHide[]>(() => {
+    if (!search) return categories
 
-  const suggestedOptions =
-    bankTransaction?.categorization_flow?.type
-    === CategorizationType.ASK_FROM_SUGGESTIONS
-      ? [
-          {
-            label: 'Suggestions',
-            options: bankTransaction.categorization_flow.suggestions.map(x =>
-              mapCategoryToOption(x),
-            ),
-          } satisfies GroupBase<CategoryOption>,
-      ]
-      : []
+    return filterCategories(categories, search.toLowerCase())
+  }, [search, categories])
 
-  const categoryOptions = flattenCategories(categories)
+  const placeholder = matches && matches.length > 1
+    ? pluralize('possible match', matches.length, true)
+    : 'Categorize or match...'
 
-  const options = [
-    ...matchOptions,
-    ...suggestedOptions,
-    ...allCategoriesDivider,
-    ...categoryOptions,
-  ]
-
-  const selected = value
-    ? value
-    : !excludeMatches
-      && matchOptions?.length === 1
-      && matchOptions[0].options.length === 1
-      ? matchOptions[0].options[0]
-      : undefined
-
-  const placeholder =
-    matchOptions?.length === 1 && matchOptions[0].options.length > 1
-      ? `${matchOptions[0].options.length} possible matches...`
-      : 'Categorize or match...'
+  const onSelect = (newValue: CategoryOption) => {
+    setIsOpen(false)
+    onChange(newValue)
+  }
 
   if (asDrawer) {
     return (
@@ -327,50 +143,91 @@ export const CategorySelect = ({
     )
   }
 
-  // The menu does not show in all cases unless the
-  // menuPortalTarget and styles lines exist
-  // See: https://stackoverflow.com/questions/55830799/how-to-change-zindex-in-react-select-drowpdown
   return (
-    <Select<CategoryOption>
-      name={name}
-      className={`Layer__category-menu Layer__select ${className ?? ''}`}
-      classNamePrefix='Layer__select'
-      classNames={{
-        menu: () => 'Layer__select__menu--lg',
-      }}
-      options={options}
-      isSearchable={true}
-      placeholder={placeholder}
-      defaultValue={selected}
-      formatOptionLabel={props => (
-        <div className='Layer__select__option-label'>
-          {props.type === 'match' && (
-            <Badge size={BadgeSize.SMALL} icon={<MinimizeTwo size={11} />}>
-              Match
-            </Badge>
-          )}
-          <span>{props.payload.display_name}</span>
-        </div>
-      )}
-      value={value}
-      onChange={newValue => newValue && onChange(newValue)}
-      getOptionLabel={category => category.payload.display_name}
-      getOptionValue={category => category.payload.id}
-      menuPortalTarget={document.body}
-      styles={{
-        menuPortal: base => ({ ...base, zIndex: 9999 }),
-      }}
-      components={{
-        DropdownIndicator,
-        GroupHeading,
-        Option: optionProps => (
-          <Option {...optionProps} showTooltips={showTooltips} />
-        ),
-      }}
-      isDisabled={disabled}
-      isOptionDisabled={option => option.disabled ?? false}
-      isOptionSelected={option =>
-        selected?.payload.display_name == option.payload.display_name}
-    />
+    <div className={classNames('Layer__category-select', className)}>
+      <DialogTrigger isOpen={isOpen} onOpenChange={v => setIsOpen(v)}>
+        <Button aria-label='Menu' className='Layer__category-select__trigger-btn' isDisabled={disabled}>
+          <Text ellipsis>{value?.payload.display_name ?? placeholder}</Text>
+          <ChevronDown size={16} />
+        </Button>
+        <input type='hidden' name={name} value={value?.payload.display_name} />
+        <Popover className='Layer__category-select__popover'>
+          <Dialog>
+            <Input
+              placeholder='Search category'
+              className='Layer__category-select__search-input'
+              value={search}
+              onChange={e => setSearch((e.target as HTMLInputElement).value)}
+            />
+            {(matches && matches.length > 0)
+            && (
+              <MenuSection>
+                <ListBox slot='listbox' aria-label='Matches'>
+                  <ListBoxSection>
+                    <Header slot='header'>
+                      <Text size={TextSize.xs} status='disabled'>Match</Text>
+                    </Header>
+                    {matches?.map((option, index) => {
+                      /** @TODO - account name - need to find in all categories */
+                      return (
+                        <ListBoxItem className='Layer__category-select__ms-list-item' key={index} textValue={option.payload.display_name} onAction={() => onSelect(option)}>
+                          <Text slot='name'>{option.payload.display_name}</Text>
+                          {/* <Text slot='account' size={TextSize.sm}>{option.payload.option_type}</Text> */}
+                        </ListBoxItem>
+                      )
+                    })}
+                  </ListBoxSection>
+                </ListBox>
+              </MenuSection>
+            )}
+
+            {(suggestions && suggestions.length > 0)
+            && (
+              <MenuSection>
+                <ListBox slot='listbox' aria-label='Suggestions'>
+                  <ListBoxSection>
+                    <Header slot='header'>
+                      <Text size={TextSize.xs} status='disabled'>Suggestions</Text>
+                    </Header>
+                    {suggestions.map((option, index) => {
+                      /** @TODO - account name - need to find in all categories */
+                      return (
+                        <ListBoxItem className='Layer__category-select__ms-list-item' key={index} textValue={option.payload.display_name} onAction={() => onSelect(option)}>
+                          <Text slot='name'>{option.payload.display_name}</Text>
+                          {/* <Text slot='account' size={TextSize.sm}>{option.payload.option_type}</Text> */}
+                        </ListBoxItem>
+                      )
+                    })}
+                  </ListBoxSection>
+                </ListBox>
+              </MenuSection>
+            )}
+
+            <MenuSection>
+              <ListBox
+                items={allCategories as unknown as CategoryWithHide[]}
+                className='Layer__category-select__all-categories'
+                aria-label='All categories'
+              >
+                {section => (
+                  <ListBoxSection
+                    className='Layer__category-select__list-box-section'
+                    id={`${section.type}-${'id' in section ? section.id : section.stable_name}`}
+                    style={{
+                      display: section.hide ? 'none' : 'flex',
+                    }}
+                  >
+                    <CategoryList option={section} level={0} accountName={section.display_name} onSelect={onSelect} selected={value} />
+                  </ListBoxSection>
+                )}
+              </ListBox>
+            </MenuSection>
+
+            <MenuItem text='Add new category' icon={<Plus size={11} />} />
+
+          </Dialog>
+        </Popover>
+      </DialogTrigger>
+    </div>
   )
 }
