+++ src/components/CategorySelect/categorySelect.scss
@@ -1,0 +1,130 @@
+.Layer__category-select {
+  display: flex;
+  flex: 1;
+}
+
+.Layer__category-select__trigger-btn {
+  background-color: var(--color-base-0);
+  border: 1px solid var(--color-base-100);
+  padding: var(--spacing-xs) var(--spacing-sm);
+  border-radius: var(--border-radius-3xs);
+  cursor: pointer;
+  flex: 1;
+  display: inline-flex;
+  align-items: center;
+  justify-content: space-between;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  white-space: nowrap;
+
+  &:focus-visible {
+    outline-color: var(--color-base-600);
+  }
+}
+
+.Layer__category-select__popover {
+  background: var(--color-base-0);
+  border-radius: var(--border-radius-xs);
+  box-shadow: 0 0 0 1px rgb(25 25 25 / 6%), 0 4px 12px 0 var(--base-transparent-1), 0 2px 4px 0 rgb(25 25 25 / 6%);
+
+  [slot='listbox'] [slot='header'] {
+    color: var(--color-base-600);
+    font-size: 12px;
+    padding: var(--spacing-xs) var(--spacing-lg);
+  }
+
+  & > [role='dialog']:focus-visible {
+    outline-color: var(--color-base-300);
+  }
+}
+
+.Layer__category-select__search-input {
+  width: 100%;
+  background-color: var(--color-base-50);
+  padding: var(--spacing-sm) var(--spacing-lg);
+  border-bottom-left-radius: 0;
+  border-bottom-right-radius: 0;
+  border-width: 0;
+  margin: 0;
+}
+
+.Layer__category-select__all-categories {
+  max-height: 30dvh;
+  overflow: auto;
+}
+
+.Layer__category-select__menu-section {
+  padding: var(--spacing-sm) 0;
+  border-bottom: 1px solid var(--color-base-200);
+}
+
+.Layer__category-select__menu-item {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: var(--spacing-xs);
+  padding: var(--spacing-sm) var(--spacing-lg);
+  cursor: pointer;
+}
+
+.Layer__category-select__ms-list-item {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: var(--spacing-xs);
+  padding: var(--spacing-3xs) var(--spacing-lg);
+  cursor: pointer;
+  transition: background-color 0.2s ease-in-out;
+
+  &:hover {
+    background-color: var(--color-base-50);
+  }
+
+  &:focus-within {
+    background-color: var(--color-base-50);
+    outline: none;
+  }
+
+  header {
+    font-size: 12px;
+    opacity: 0.4;
+  }
+}
+
+.Layer__category-select__list-box-section {
+  display: flex;
+  width: 100%;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: space-between;
+  padding: 0 0 var(--spacing-xs) 0;
+  gap: 0;
+
+  header {
+    padding-top: var(--spacing-xs);
+    padding-bottom: var(--spacing-3xs);
+  }
+}
+
+.Layer__category-select__list-item {
+  display: flex;
+  box-sizing: border-box;
+  flex-direction: column;
+  align-items: flex-start;
+  width: 100%;
+  justify-content: space-between;
+  gap: var(--spacing-xs);
+  padding: var(--spacing-3xs) var(--spacing-lg);
+  cursor: pointer;
+  transition: background-color 0.2s ease-in-out;
+
+  &:hover {
+    background-color: var(--color-base-50);
+  }
+
+  &:focus-within {
+    background-color: var(--color-base-50);
+    outline: none;
+  }
+}
