--- src/components/CategorySelect2/categorySelect.scss
@@ -1,4 +1,0 @@
-.Layer__category-select__popover {
-  background: var(--color-base-50);
-  border-radius: var(--border-radius-xs);
-}
