+++ src/styles/index.scss
@@ -40,7 +40,7 @@
 @forward '../components/ProfitAndLossHeader/profitAndLossHeader';
 @forward '../components/BookkeepingStatus/bookkeepingStatus';
 @forward '../components/BookkeepingStatus/BookkeepingStatusReportRow';
-@forward '../components/CategorySelect2/categorySelect';
+@forward '../components/CategorySelect/categorySelect';
 @forward '../components/Tabs/tabs';
 @forward '../components/MobilePanel/mobilePanel';
 @forward './ledger_account';
