+++ src/styles/typography.scss
@@ -57,6 +57,10 @@
   margin: 0;
 }
 
+.Layer__text--xs {
+  font-size: var(--text-xs);
+}
+
 .Layer__text--sm {
   font-size: var(--text-sm);
 }
